Refactor TaskItem even further
