+++ todo-app/src/Components/TaskExpander.css
@@ -1,0 +1,35 @@
+.toDoOptions {
+  grid-row: 1;
+  grid-column: 3;
+  border-left: 1px solid rgb(219, 219, 219);
+}
+
+.toDoDropdown {
+  color: rgb(48, 48, 48);
+  font-size: 21px;
+  text-align: center;
+  border-top-right-radius: 10px;
+  border-bottom-right-radius: 10px;
+  background-color: rgb(248, 248, 248);
+  box-sizing: border-box;
+  height: 100%;
+  transition-property: all;
+  transition-duration: 0.2s;
+  width: 100%;
+  padding-top: 4px;
+}
+
+.optionsArrow {
+  transition-property: all;
+  transition-duration: 0.2s;
+}
+
+.toDoDropdown.selected {
+  background-color: rgb(49, 49, 209);
+  color: white;
+  border-bottom-right-radius: 0;
+}
+
+.rotated {
+  transform: rotate(90deg);
+}
