+++ todo-app/src/Components/TaskItem.css
@@ -17,82 +17,13 @@
   opacity: 0;
 }
 
-.toDoCheckbox {
-  display: flex;
-  grid-row: 1;
-  grid-column: 1;
-  padding: 5px;
-}
-
-.toDoLabel {
-  display: flex;
-  grid-row: 1;
-  grid-column: 2;
-  font-size: 18px;
-  box-sizing: border-box;
-}
-
-.toDoLabel > input {
-  border: 1px solid rgb(221, 221, 221);
-  border-radius: 5px;
-  width: 100%;
-  padding: 5px;
-  font-family: 'Roboto Slab', Arial, Helvetica, sans-serif;
-  font-size: 18px;
-}
-
 .toDoItem input[type='checkbox'] {
   height: 18px;
   width: 18px;
 }
 
-.toDoOptions {
-  grid-row: 1;
-  grid-column: 3;
-  border-left: 1px solid rgb(219, 219, 219);
-}
-
-.toDoDropdown {
-  color: rgb(48, 48, 48);
-  font-size: 21px;
-  text-align: center;
-  border-top-right-radius: 10px;
-  border-bottom-right-radius: 10px;
-  background-color: rgb(248, 248, 248);
-  box-sizing: border-box;
-  height: 100%;
-  transition-property: all;
-  transition-duration: 0.2s;
-  width: 100%;
-  padding-top: 4px;
-}
-
-.optionsArrow {
-  transition-property: all;
-  transition-duration: 0.2s;
-}
-
-.toDoDropdown.selected {
-  background-color: rgb(49, 49, 209);
-  color: white;
-  border-bottom-right-radius: 0;
-}
-
-.strikethrough {
-  text-decoration: line-through;
-  color: rgb(148, 147, 147)
-}
-
-.rotated {
-  transform: rotate(90deg);
-}
-
 .grayText {
   color: rgb(175, 175, 175);
-}
-
-.hidden {
-  display: none;
 }
 
 .priorityMarker {
