+++ todo-app/src/Components/TaskItem.js
@@ -1,12 +1,14 @@
 import { useEffect, useRef, useState } from 'react';
 import TaskItemOptions from './TaskItemOptions';
+import TaskCheckbox from './TaskCheckbox';
 import './TaskItem.css';
+import TaskTextLabel from './TaskTextLabel';
+import TaskExpander from './TaskExpander';
 
 export default function TaskItem(props) {
   const [shouldFadeOut, setShouldFadeOut] = useState(false);
 
   const textInput = useRef(null);
-  const TEXT_CHAR_LIMIT = 100;
   const DISAPPEAR_DURATION_MS = 500;
 
   useEffect(() => {
@@ -14,10 +16,6 @@ export default function TaskItem(props) {
       textInput.current.focus();
     }
   }, []);
-
-  function elideText(text) {
-    return text.length > TEXT_CHAR_LIMIT ? text.slice(0, TEXT_CHAR_LIMIT) + '...' : text;
-  }
 
   function priorityToMarker(priority) {
     return <div className={`priorityExclamationDiv priority${priority}`}>{'!'.repeat(priority)}</div>;
@@ -61,41 +59,15 @@ export default function TaskItem(props) {
         {priorityToMarker(props.priority)}
       </div>
       <div className={`toDoItem supportsInvisibility ${shouldFadeOut ? 'invisible' : ''}`}>
-          <div className="toDoCheckbox">
-            <input  type="checkbox" 
-                    id={`label-${props.id}`} 
-                    checked={props.isCompleted} 
-                    onChange={e => handleCompletion(e)}/>
-          </div>
-          <div className="toDoLabel">
-            {
-              props.taskInEditModeId !== props.id &&
-                <label id={`label-${props.id}`} className={props.isCompleted || shouldFadeOut ? 'strikethrough' : ''}>
-                  {
-                    props.expandedTaskId === props.id
-                    ? props.taskName
-                    : elideText(props.taskName)
-                  }
-                </label>
-            }
-            <input 
-              value={props.taskName} 
-              onChange={e => props.setTaskProperty(props.id, 'taskName', e.target.value)}
-              onKeyUp={e => {if (e.key === 'Enter') props.setTaskInEditModeId(null)}} 
-              ref={textInput}
-              className={props.taskInEditModeId !== props.id ? 'hidden' : ''}
-              onBlur={() => deselectOnEditMode()} />
-          </div>
+          <TaskCheckbox id={props.id} checked={props.isCompleted} handleCompletion={handleCompletion} />
+          <TaskTextLabel 
+            {...props}
+            textInput={textInput}
+            deselectOnEditMode={deselectOnEditMode}
+            shouldFadeOut={shouldFadeOut} />
           {
             !props.isCompleted && 
-              <div className="toDoOptions">
-                <div  className={`toDoDropdown ${props.expandedTaskId === props.id ? 'selected' : ''}`}
-                      onClick={() => props.expandTaskCallback(props.id)}>
-                  <div className={`optionsArrow ${props.expandedTaskId === props.id ? 'rotated' : ''}`}>
-                    ➔
-                  </div>
-                </div>
-              </div>
+              <TaskExpander {...props} />
           }
           {
             props.expandedTaskId === props.id &&
