+++ todo-app/src/Components/TaskTextLabel.css
@@ -1,0 +1,25 @@
+.toDoLabel {
+  display: flex;
+  grid-row: 1;
+  grid-column: 2;
+  font-size: 18px;
+  box-sizing: border-box;
+}
+
+.toDoLabel > input {
+  border: 1px solid rgb(221, 221, 221);
+  border-radius: 5px;
+  width: 100%;
+  padding: 5px;
+  font-family: 'Roboto Slab', Arial, Helvetica, sans-serif;
+  font-size: 18px;
+}
+
+.strikethrough {
+  text-decoration: line-through;
+  color: rgb(148, 147, 147)
+}
+
+.hidden {
+  display: none;
+}
